buttonDemo
This is a basic application using basic elements. Using this concept you can create unique applications to showcase your creativity. In this project, we are just using buttons, and when we click that buttons the text will be toasted like a popup. It gives result or information about that button in this code .you can probably give whatever text you want to display in the toast method associated with a particular button. You will also understand layout modelling and spacing from this project. Think about a simple application similar to this and proceed.